refactor: 컨버터 수정 및 넘버 유틸 패키지 변경 (#410)
