--- backend/src/main/java/wooteco/prolog/member/ui/LocalDateConverter.java
@@ -1,15 +1,0 @@
-package wooteco.prolog.member.ui;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class LocalDateConverter implements Converter<String, LocalDate> {
-
-    @Override
-    public LocalDate convert(String source) {
-        return LocalDate.parse(source, DateTimeFormatter.ofPattern("yyyyMMdd"));
-    }
-}
