--- backend/src/main/java/wooteco/prolog/member/ui/MonthConverter.java
@@ -1,14 +1,0 @@
-package wooteco.prolog.member.ui;
-
-import java.time.Month;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MonthConverter implements Converter<String, Month> {
-
-    @Override
-    public Month convert(String source) {
-        return Month.of(Integer.parseInt(source));
-    }
-}
